REFACTOR: Split out automerge_util from automerge_json
