+++ packages/backend/src/document.rs
@@ -4,9 +4,8 @@ use crate::app::{AppCtx, AppError, AppState};
 use crate::autosave::ensure_autosave_listener;
 use crate::user_state_updates::{update_ref_for_users, update_user_state};
 use chrono::{DateTime, Utc};
-use notebook_types::automerge_json::{
-    copy_doc_at_heads, hydrate_to_json, populate_automerge_from_json,
-};
+use notebook_types::automerge_json::{hydrate_to_json, populate_automerge_from_json};
+use notebook_types::automerge_util::copy_doc_at_heads;
 use samod::DocumentId;
 use serde_json::Value;
 use uuid::Uuid;
